+++ lightly_studio/src/lightly_studio/resolvers/collection_resolver/__init__.py
@@ -4,6 +4,10 @@ from lightly_studio.resolvers.collection_resolver.check_collection_type import (
     check_collection_type,
 )
 from lightly_studio.resolvers.collection_resolver.create import create
+from lightly_studio.resolvers.collection_resolver.create_group_components import (
+    create_group_components,
+)
+from lightly_studio.resolvers.collection_resolver.deep_copy import deep_copy
 from lightly_studio.resolvers.collection_resolver.delete import delete
 from lightly_studio.resolvers.collection_resolver.export import (
     export,
@@ -20,6 +24,9 @@ from lightly_studio.resolvers.collection_resolver.get_collection_details import 
 )
 from lightly_studio.resolvers.collection_resolver.get_collections_overview import (
     get_collections_overview,
+)
+from lightly_studio.resolvers.collection_resolver.get_group_components import (
+    get_group_components,
 )
 from lightly_studio.resolvers.collection_resolver.get_hierarchy import (
     get_hierarchy,
@@ -38,6 +45,8 @@ from lightly_studio.resolvers.collection_resolver.update import update
 __all__ = [
     "check_collection_type",
     "create",
+    "create_group_components",
+    "deep_copy",
     "delete",
     "export",
     "get_all",
@@ -47,6 +56,7 @@ __all__ = [
     "get_collections_overview",
     "get_dataset",
     "get_filtered_samples_count",
+    "get_group_components",
     "get_hierarchy",
     "get_or_create_child_collection",
     "get_parent_collection_by_sample_id",