+++ lightly_studio_view/src/lib/components/Header/Header.svelte
@@ -12,6 +12,8 @@
     import Menu from '$lib/components/Header/Menu.svelte';
     import type { CollectionView } from '$lib/api/lightly_studio_local';
     import { useRootCollectionOptions } from '$lib/hooks/useRootCollection/useRootCollection';
+    import UserAvatar from '$lib/components/UserAvatar/UserAvatar.svelte';
+    import useAuth from '$lib/hooks/useAuth/useAuth';
 
     let { collection }: { collection: CollectionView } = $props();
 
@@ -19,7 +21,7 @@
     const { settingsStore } = useSettings();
 
     const hasEmbeddingsQuery = useHasEmbeddings({ collectionId: collection.collection_id });
-    const hasEmbeddings = $derived(!!hasEmbeddingsQuery.data);
+    const hasEmbeddings = $derived(!!$hasEmbeddingsQuery.data);
 
     const { rootCollection } = useRootCollectionOptions({ collectionId: collection.collection_id });
 
@@ -43,6 +45,8 @@
             await executeReversibleAction(latestAction.id);
         }
     };
+
+    const { user } = useAuth();
 </script>
 
 <svelte:window onkeydown={handleKeyDown} />
@@ -94,6 +98,11 @@
                         <Pencil class="size-4" />
                         <span>Edit Annotations</span>
                     </Button>
+                {/if}
+                {#if user}
+                    <div data-testid="header-user-avatar">
+                        <UserAvatar {user} />
+                    </div>
                 {/if}
             </div>
         </div>