+++ lightly_studio_view/src/lib/components/SampleDetails/SampleInstanceSegmentationRect/SampleInstanceSegmentationRect.svelte
@@ -54,11 +54,10 @@
         collectionId,
         sampleId,
         sample,
-        labels: $labels.data ?? [],
         refetch
     });
 
-    const annotationLabelContext = useAnnotationLabelContext();
+    const { context: annotationLabelContext, setIsDrawing } = useAnnotationLabelContext();
 
     let brushPath = $state<{ x: number; y: number }[]>([]);
     let workingMask = $state<Uint8Array | null>(null);
@@ -66,7 +65,7 @@
     let selectedAnnotation = $state<AnnotationView | null>(null);
 
     $effect(() => {
-        if (annotationLabelContext.isDrawing) return;
+        setIsDrawing(false);
 
         if (!annotationLabelContext.annotationId) {
             previewRLE = [];
@@ -131,14 +130,24 @@
         updatePreview();
     }}
     onpointerleave={() =>
-        finishBrush(workingMask, selectedAnnotation, annotationApi?.updateAnnotation)}
+        finishBrush(
+            workingMask,
+            selectedAnnotation,
+            $labels.data ?? [],
+            annotationApi?.updateAnnotation
+        )}
     onpointerup={() =>
-        finishBrush(workingMask, selectedAnnotation, annotationApi?.updateAnnotation)}
+        finishBrush(
+            workingMask,
+            selectedAnnotation,
+            $labels.data ?? [],
+            annotationApi?.updateAnnotation
+        )}
     onpointerdown={(e) => {
         const point = getImageCoordsFromMouse(e, interactionRect, sample.width, sample.height);
         if (!point) return;
 
-        annotationLabelContext.isDrawing = true;
+        setIsDrawing(true);
         brushPath.push(point);
 
         if (!workingMask) {