+++ lightly_studio_view/src/lib/hooks/useInstanceSegmentationBrush.ts
@@ -13,33 +13,37 @@ export function useInstanceSegmentationBrush({
     collectionId,
     sampleId,
     sample,
-    refetch,
-    labels
+    refetch
 }: {
     collectionId: string;
     sampleId: string;
     sample: { width: number; height: number };
-    labels: {
-        annotation_label_id?: string;
-        annotation_label_name?: string;
-    }[];
     refetch: () => void;
 }) {
     const { createLabel } = useCreateLabel({ collectionId });
     const { createAnnotation } = useCreateAnnotation({ collectionId });
-    const annotationLabelContext = useAnnotationLabelContext();
+    const {
+        context: annotationLabelContext,
+        setIsDrawing,
+        setAnnotationLabel,
+        setLastCreatedAnnotationId,
+        setAnnotationId
+    } = useAnnotationLabelContext();
 
     const finishBrush = async (
         workingMask: Uint8Array | null,
         selectedAnnotation: AnnotationView | null,
+        labels: {
+            annotation_label_id?: string;
+            annotation_label_name?: string;
+        }[],
         updateAnnotation?: (input: AnnotationUpdateInput) => Promise<void>
     ) => {
         if (!annotationLabelContext.isDrawing || !workingMask) {
-            annotationLabelContext.isDrawing = false;
+            setIsDrawing(false);
+
             return;
         }
-
-        annotationLabelContext.isDrawing = false;
 
         const bbox: BoundingBox | null = computeBoundingBoxFromMask(
             workingMask,
@@ -53,16 +57,14 @@ export function useInstanceSegmentationBrush({
         }
 
         const rle = encodeBinaryMaskToRLE(workingMask);
-
         if (selectedAnnotation) {
             try {
                 await updateAnnotation?.({
-                    annotation_id: annotationLabelContext.annotationId!,
+                    annotation_id: selectedAnnotation.sample_id,
                     collection_id: collectionId,
                     bounding_box: bbox,
                     segmentation_mask: rle
                 });
-
                 refetch();
                 return;
             } catch (error) {
@@ -74,12 +76,12 @@ export function useInstanceSegmentationBrush({
         let label =
             labels?.find(
                 (l) => l.annotation_label_name === annotationLabelContext.annotationLabel
-            ) ?? labels?.find((l) => l.annotation_label_name === 'default');
+            ) ?? labels?.find((l) => l.annotation_label_name === 'DEFAULT');
 
         if (!label) {
             label = await createLabel({
                 dataset_id: collectionId,
-                annotation_label_name: 'default'
+                annotation_label_name: 'DEFAULT'
             });
         }
 
@@ -94,9 +96,9 @@ export function useInstanceSegmentationBrush({
             annotation_label_id: label.annotation_label_id!
         });
 
-        annotationLabelContext.annotationLabel = label.annotation_label_name;
-        annotationLabelContext.annotationId = newAnnotation.sample_id;
-        annotationLabelContext.lastCreatedAnnotationId = newAnnotation.sample_id;
+        setAnnotationLabel(label.annotation_label_name!);
+        setAnnotationId(newAnnotation.sample_id);
+        setLastCreatedAnnotationId(newAnnotation.sample_id);
 
         refetch();
     };